refactor: update import declarations, mode enums to typings/enums
