+++ packages/api/src/models/room/index.ts
@@ -1,3 +1,2 @@
 export * from "./room";
-export * from "./enums";
 export * from "./__tests__/room.stub";
